requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.1
Werkzeug==2.2.2
